Refactors linting logic and updates module structure
Reorganizes linting functionality into a new directory for better modularity
Simplifies and removes redundant code from lint-selection implementation
Updates import paths to reflect the new module structure
